+++ tests/unit/test_main.py
@@ -8,11 +8,7 @@ from kytos.lib.helpers import (
     get_kytos_event_mock,
     get_switch_mock,
 )
-from napps.amlight.flow_stats.main import GenericFlow, Main
-from napps.kytos.of_core.v0x04.flow import Action as Action40
-from napps.kytos.of_core.v0x04.match_fields import MatchFieldFactory
-from napps.kytos.of_core.v0x04.flow import Match as Match40
-from pyof.foundation.basic_types import UBInt32
+from napps.amlight.flow_stats.main import Main
 
 
 # pylint: disable=too-many-public-methods, too-many-lines
@@ -93,129 +89,193 @@ class TestMain(TestCase):
             ),
         ]
         urls = self.get_napp_urls(self.napp)
-        self.assertEqual(len(expected_urls), len(urls))
+        assert len(expected_urls) == len(urls)
 
     def test_packet_count__fail(self):
-        """Test packet_count rest call with wrong flow_id."""
+        """Test bytes_count rest call with wrong flow_id."""
         flow_id = "123456789"
         rest_name = "packet_count"
         response = self._get_rest_response(rest_name, flow_id)
 
-        self.assertEqual(response.data, b"Flow does not exist")
+        assert response.data == b"Flow does not exist"
 
-    def test_packet_count(self):
+    @patch("napps.amlight.flow_stats.main.Main.flow_from_id")
+    def test_packet_count(self, mock_from_flow):
         """Test packet_count rest call."""
-        flow_id = "1"
+        flow_id = '1'
+        dpid_id = '1'
+        mock_from_flow.return_value = self._get_mocked_flow_base()
+
         rest_name = "packet_count"
         self._patch_switch_flow(flow_id)
-        response = self._get_rest_response(rest_name, flow_id)
 
+        response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
-        self.assertEqual(json_response["flow_id"], flow_id)
-        self.assertEqual(json_response["packet_counter"], 40)
-        self.assertEqual(json_response["packet_per_second"], 2.0)
+        assert json_response["flow_id"] == flow_id
+        assert json_response["packet_counter"] == 40
+        assert json_response["packet_per_second"] == 2.0
 
-    def test_bytes_count_fail(self):
+    def test_bytes_count__fail(self):
         """Test bytes_count rest call with wrong flow_id."""
         flow_id = "123456789"
         rest_name = "bytes_count"
         response = self._get_rest_response(rest_name, flow_id)
 
-        self.assertEqual(response.data, b"Flow does not exist")
+        assert response.data == b"Flow does not exist"
 
-    def test_bytes_count(self):
+    @patch("napps.amlight.flow_stats.main.Main.flow_from_id")
+    def test_bytes_count(self, mock_from_flow):
         """Test bytes_count rest call."""
-        flow_id = "1"
+        flow_id = '1'
+        dpid_id = '1'
+        mock_from_flow.return_value = self._get_mocked_flow_base()
+
         rest_name = "bytes_count"
         self._patch_switch_flow(flow_id)
-        response = self._get_rest_response(rest_name, flow_id)
 
+        response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
-        self.assertEqual(json_response["flow_id"], flow_id)
-        self.assertEqual(json_response["bytes_counter"], 10)
-        self.assertEqual(json_response["bits_per_second"], 4.0)
+        assert json_response["flow_id"] == flow_id
+        assert json_response["bytes_counter"] == 10
+        assert json_response["bits_per_second"] == 4.0
 
-    def test_packet_count_per_flow(self):
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_packet_count_per_flow(self, mock_from_flow):
         """Test packet_count_per_flow rest call."""
-        flow_id = "1"
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
         rest_name = "packet_count/per_flow"
         self._patch_switch_flow(flow_id)
 
-        dpid_id = 111
+        mock_from_flow.return_value = flow_by_sw
         response = self._get_rest_response(rest_name, dpid_id)
-
         json_response = json.loads(response.data)
-        self.assertEqual(json_response[0]["flow_id"], flow_id)
-        self.assertEqual(json_response[0]["packet_counter"], 40)
-        self.assertEqual(json_response[0]["packet_per_second"], 2.0)
+        assert json_response[0]["flow_id"] == flow_id
+        assert json_response[0]["packet_counter"] == 40
+        assert json_response[0]["packet_per_second"] == 2.0
 
-    def test_packet_count_sum(self):
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_packet_count_sum(self, mock_from_flow):
         """Test packet_count_sum rest call."""
-        flow_id = "1"
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
         rest_name = "packet_count/sum"
         self._patch_switch_flow(flow_id)
 
-        dpid_id = 111
+        mock_from_flow.return_value = flow_by_sw
         response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
+        assert json_response == 40
 
-        self.assertEqual(json_response, 40)
-
-    def test_bytes_count_per_flow(self):
-        """Test bytes_count_per_flow rest call."""
-        flow_id = "1"
-        rest_name = "bytes_count/per_flow"
-        self._patch_switch_flow(flow_id)
-
-        dpid_id = 111
-        response = self._get_rest_response(rest_name, dpid_id)
-
-        json_response = json.loads(response.data)
-        self.assertEqual(json_response[0]["flow_id"], flow_id)
-        self.assertEqual(json_response[0]["bytes_counter"], 10)
-        self.assertEqual(json_response[0]["bits_per_second"], 4.0)
-
-    def test_bytes_count_sum(self):
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_bytes_count_sum(self, mock_from_flow):
         """Test bytes_count_sum rest call."""
-        flow_id = "1"
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
         rest_name = "bytes_count/sum"
         self._patch_switch_flow(flow_id)
 
-        dpid_id = 111
+        mock_from_flow.return_value = flow_by_sw
         response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
+        assert json_response == 10
 
-        self.assertEqual(json_response, 10)
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_bytes_count_per_flow(self, mock_from_flow):
+        """Test bytes_count_per_flow rest call."""
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
 
-    @patch("napps.amlight.flow_stats.main.Main.match_flows")
-    def test_flow_stats(self, mock_match_flows):
-        """Test flow_match rest call."""
-        flow = GenericFlow()
-        flow.actions = [
-            Action40.from_dict(
-                {
-                    "action_type": "output",
-                    "port": "1",
-                }
-            ),
-        ]
-        flow.version = "0x04"
-        mock_match_flows.return_value = [flow]
-
-        flow_id = "1"
-        rest_name = "flow/stats"
+        rest_name = "bytes_count/per_flow"
         self._patch_switch_flow(flow_id)
 
-        dpid_id = "aa:00:00:00:00:00:00:11"
+        mock_from_flow.return_value = flow_by_sw
         response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
+        assert json_response[0]["flow_id"] == flow_id
+        assert json_response[0]["bytes_counter"] == 10
+        assert json_response[0]["bits_per_second"] == 4.0
 
-        self.assertEqual(response.status_code, 200)
-        print(json_response)
-        self.assertEqual(json_response[0]["actions"][0]["action_type"],
-                         "output")
-        self.assertEqual(json_response[0]["actions"][0]["port"], "1")
-        self.assertEqual(json_response[0]["version"], "0x04")
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_flow_stats(self, mock_from_flow):
+        """Test flow_stats rest call."""
+        flow_stats = {
+            'byte_count': 148,
+            'duration_sec': 1589,
+            'duration_nsec': 556000000,
+            'packet_count': 2
+            }
+        flow_stats_dict_mock = {'6055f13593fad45e0b4699f49d56b105': flow_stats}
+        flow_by_sw = {"00:00:00:00:00:00:00:01": flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
+        api = get_test_client(self.napp.controller, self.napp)
+        url = f"{self.server_name_url}/flow/stats?dpid=00:00:00:00:00:00:00:01"
+
+        response = api.get(url)
+        expected = flow_by_sw
+        assert response.json == expected
+        assert response.status_code == 200
+
+    @patch("napps.amlight.flow_stats.main.Main._flow_stats")
+    def test_flow_stats_with_dpid(self, mock_from_flow):
+        """Test flow_stats rest call."""
+        flow_stats = {
+            'byte_count': 148,
+            'duration_sec': 1589,
+            'duration_nsec': 556000000,
+            'packet_count': 2
+            }
+        flow_stats_dict_mock = {'6055f13593fad45e0b4699f49d56b105': flow_stats}
+        flow_by_sw = {"00:00:00:00:00:00:00:01": flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
+        api = get_test_client(self.napp.controller, self.napp)
+        url = f"{self.server_name_url}/flow/stats?dpid=00:00:00:00:00:00:00:01"
+
+        response = api.get(url)
+        expected = flow_by_sw
+        assert response.json == expected
+        assert response.status_code == 200
 
     def _patch_switch_flow(self, flow_id):
         """Helper method to patch controller to return switch/flow data."""
@@ -223,7 +283,6 @@ class TestMain(TestCase):
         flow = self._get_mocked_flow_stats()
         flow.id = flow_id
         switch = MagicMock()
-        switch.generic_flows = [flow]
         self.napp.controller.switches = {"1": switch}
         self.napp.controller.get_switch_by_dpid = MagicMock()
         self.napp.controller.get_switch_by_dpid.return_value = switch
@@ -299,13 +358,10 @@ class TestMain(TestCase):
         self.napp.handle_stats_received(event)
         mock_handle_stats.assert_not_called()
 
-    @patch("napps.amlight.flow_stats.main.GenericFlow.from_replies_flows")
-    def test_handle_stats_reply_received(self, mock_from_flow):
+    def test_handle_stats_reply_received(self):
         """Test handle_stats_reply_received call."""
-        mock_from_flow.return_value = self._get_mocked_flow_base()
 
-        event_switch = MagicMock()
         flows_mock = self._get_mocked_multipart_replies_flows()
-        self.napp.handle_stats_reply_received(event_switch, flows_mock)
+        self.napp.handle_stats_reply_received(flows_mock)
 
-        self.assertEqual(event_switch.generic_flows[0].id, 456)
+        assert list(self.napp.flows_stats_dict.values())[0].id == 456